requirements.txt
pyobjc-core
pyobjc-framework-Cocoa
pyobjc-framework-AVFoundation
pyobjc-framework-Vision
pyobjc-framework-Quartz
zxing-cpp
